server/requirements.txt
tornado==4.4.1

psycopg2==2.6.2
SQLAlchemy==1.0.14
aiopg==0.10.0

websocket-client==0.37.0



